build: 데이터바인딩 제외 코드리뷰 반영
