+++ app/src/main/res/layout/item_soso_pick.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <data>
-
+        
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -23,7 +23,6 @@
 
         <TextView
             android:id="@+id/tv_explore_soso_pick_novel_title"
-            style="@style/label1"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_marginTop="8dp"
@@ -31,6 +30,7 @@
             android:gravity="top|start"
             android:maxLines="2"
             android:minLines="2"
+            android:textAppearance="@style/label1"
             android:textColor="@color/black"
             android:textColorHint="@color/black"
             app:layout_constraintEnd_toEndOf="parent"
